Refactoring of tctl
This commit prepares TCTL for easier modifications for 2.3
Specifically:

1. CLI commands have been placed in separate go files.
2. Removed some dead code. We had commands that haven't even been hooked
   to CLI!
3. Added comments in a few places.
