+++ tool/tctl/common/tctl.go
@@ -17,114 +17,28 @@ limitations under the License.
 package common
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"net"
-	"net/url"
 	"os"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/lib/auth"
-	"github.com/gravitational/teleport/lib/auth/native"
-	"github.com/gravitational/teleport/lib/backend"
 	"github.com/gravitational/teleport/lib/client"
 	"github.com/gravitational/teleport/lib/config"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/service"
-	"github.com/gravitational/teleport/lib/services"
-	"github.com/gravitational/teleport/lib/sshutils"
 	"github.com/gravitational/teleport/lib/utils"
-	"github.com/gravitational/teleport/lib/web"
 
 	"github.com/Sirupsen/logrus"
 	"github.com/buger/goterm"
-	"github.com/gravitational/kingpin"
 	"github.com/gravitational/trace"
 	"golang.org/x/crypto/ssh"
-	kyaml "k8s.io/client-go/1.4/pkg/util/yaml"
 )
 
 type CLIConfig struct {
 	Debug        bool
 	ConfigFile   string
 	ConfigString string
-}
-
-type UserCommand struct {
-	config        *service.Config
-	login         string
-	allowedLogins string
-	roles         string
-	identities    []string
-}
-
-type NodeCommand struct {
-	config *service.Config
-	// count is optional hidden field that will cause
-	// tctl issue count tokens and output them in JSON format
-	count int
-	// format is the output format, e.g. text or json
-	format string
-	// list of roles for the new node to assume
-	roles string
-	// TTL: duration of time during which a generated node token will
-	// be valid.
-	ttl time.Duration
-	// namespace is node namespace
-	namespace string
-}
-
-type AuthCommand struct {
-	config                     *service.Config
-	authType                   string
-	genPubPath                 string
-	genPrivPath                string
-	genUser                    string
-	genTTL                     time.Duration
-	exportAuthorityFingerprint string
-	exportPrivateKeys          bool
-	output                     string
-	outputFormat               client.IdentityFileFormat
-	compatVersion              string
-	compatibility              string
-}
-
-type SAMLCommand struct {
-	config *service.Config
-	name   string
-}
-
-type AuthServerCommand struct {
-	config *service.Config
-}
-
-type TokenCommand struct {
-	config *service.Config
-	// token argument to 'tokens del' command
-	token string
-}
-
-type GetCommand struct {
-	config      *service.Config
-	ref         services.Ref
-	format      string
-	namespace   string
-	withSecrets bool
-}
-
-type CreateCommand struct {
-	config   *service.Config
-	filename string
-}
-
-type DeleteCommand struct {
-	config *service.Config
-	ref    services.Ref
 }
 
 // Run() is the same as 'make'. It helps to share the code between different
@@ -246,7 +160,6 @@ func Run(distribution string) {
 	}
 
 	applyConfig(&ccf, cfg)
-	validateConfig(cfg)
 
 	// some commands do not need a connection to client
 	switch command {
@@ -298,12 +211,7 @@ func Run(distribution string) {
 	}
 }
 
-func Ref(s kingpin.Settings) *services.Ref {
-	r := new(services.Ref)
-	s.SetValue(r)
-	return r
-}
-
+// printHeader helper prints an ASCII table header to stdout
 func printHeader(t *goterm.Table, cols []string) {
 	dots := make([]string, len(cols))
 	for i := range dots {
@@ -311,348 +219,6 @@ func printHeader(t *goterm.Table, cols []string) {
 	}
 	fmt.Fprint(t, strings.Join(cols, "\t")+"\n")
 	fmt.Fprint(t, strings.Join(dots, "\t")+"\n")
-}
-
-// Add creates a new sign-up token and prints a token URL to stdout.
-// A user is not created until he visits the sign-up URL and completes the process
-func (u *UserCommand) Add(client *auth.TunClient) error {
-	// if no local logins were specified, default to 'login'
-	if u.allowedLogins == "" {
-		u.allowedLogins = u.login
-	}
-	user := services.UserV1{
-		Name:          u.login,
-		AllowedLogins: strings.Split(u.allowedLogins, ","),
-	}
-	token, err := client.CreateSignupToken(user)
-	if err != nil {
-		return err
-	}
-	proxies, err := client.GetProxies()
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	hostname := "teleport-proxy"
-	if len(proxies) == 0 {
-		fmt.Printf("\x1b[1mWARNING\x1b[0m: this Teleport cluster does not have any proxy servers online.\nYou need to start some to be able to login.\n\n")
-	} else {
-		hostname = proxies[0].GetHostname()
-	}
-
-	// try to auto-suggest the activation link
-	_, proxyPort, err := net.SplitHostPort(u.config.Proxy.WebAddr.Addr)
-	if err != nil {
-		proxyPort = strconv.Itoa(defaults.HTTPListenPort)
-	}
-	url := web.CreateSignupLink(net.JoinHostPort(hostname, proxyPort), token)
-	fmt.Printf("Signup token has been created and is valid for %v seconds. Share this URL with the user:\n%v\n\nNOTE: make sure '%s' is accessible!\n", defaults.MaxSignupTokenTTL.Seconds(), url, hostname)
-	return nil
-}
-
-// List prints all existing user accounts
-func (u *UserCommand) List(client *auth.TunClient) error {
-	users, err := client.GetUsers()
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	coll := &userCollection{users: users}
-	coll.writeText(os.Stdout)
-	return nil
-}
-
-// Delete deletes teleport user(s). User IDs are passed as a comma-separated
-// list in UserCommand.login
-func (u *UserCommand) Delete(client *auth.TunClient) error {
-	for _, l := range strings.Split(u.login, ",") {
-		if err := client.DeleteUser(l); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("User '%v' has been deleted\n", l)
-	}
-	return nil
-}
-
-// Update updates existing user
-func (u *UserCommand) Update(client *auth.TunClient) error {
-	user, err := client.GetUser(u.login)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	roles := strings.Split(u.roles, ",")
-	for _, role := range roles {
-		if _, err := client.GetRole(role); err != nil {
-			return trace.Wrap(err)
-		}
-	}
-	user.SetRoles(roles)
-	if err := client.UpsertUser(user); err != nil {
-		return trace.Wrap(err)
-	}
-	fmt.Printf("%v has been updated with roles %v\n", user.GetName(), strings.Join(user.GetRoles(), ","))
-	return nil
-}
-
-// Invite generates a token which can be used to add another SSH node
-// to a cluster
-func (u *NodeCommand) Invite(client *auth.TunClient) error {
-	if u.count < 1 {
-		return trace.BadParameter("count should be > 0, got %v", u.count)
-	}
-	// parse --roles flag
-	roles, err := teleport.ParseRoles(u.roles)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	var tokens []string
-	for i := 0; i < u.count; i++ {
-		token, err := client.GenerateToken(roles, u.ttl)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		tokens = append(tokens, token)
-	}
-
-	authServers, err := client.GetAuthServers()
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	if len(authServers) == 0 {
-		return trace.Errorf("This cluster does not have any auth servers running")
-	}
-
-	// output format swtich:
-	if u.format == "text" {
-		for _, token := range tokens {
-			fmt.Printf(
-				"The invite token: %v\nRun this on the new node to join the cluster:\n> teleport start --roles=%s --token=%v --auth-server=%v\n\nPlease note:\n",
-				token, strings.ToLower(roles.String()), token, authServers[0].GetAddr())
-		}
-		fmt.Printf("  - This invitation token will expire in %d minutes\n", int(u.ttl.Minutes()))
-		fmt.Printf("  - %v must be reachable from the new node, see --advertise-ip server flag\n", authServers[0].GetAddr())
-		fmt.Printf(`  - For tokens of type "trustedcluster", tctl needs to be used to create a TrustedCluster resource. See the Admin Guide for more details.`)
-	} else {
-		out, err := json.Marshal(tokens)
-		if err != nil {
-			return trace.Wrap(err, "failed to marshal tokens")
-		}
-		fmt.Printf(string(out))
-	}
-	return nil
-}
-
-// ListActive retreives the list of nodes who recently sent heartbeats to
-// to a cluster and prints it to stdout
-func (u *NodeCommand) ListActive(client *auth.TunClient) error {
-	nodes, err := client.GetNodes(u.namespace)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	coll := &serverCollection{servers: nodes}
-	coll.writeText(os.Stdout)
-	return nil
-}
-
-// ExportSAML outputs a certificate in CRT format to be used by an idP (identity
-// provider) to validate an AuthnRequest.
-func (s *SAMLCommand) ExportSAML(client *auth.TunClient) error {
-	sc, err := client.GetSAMLConnector(s.name, false)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	fmt.Printf("%v", sc.GetSigningKeyPair().Cert)
-	return nil
-}
-
-// ExportAuthorities outputs the list of authorities in OpenSSH compatible formats
-// If --type flag is given, only prints keys for CAs of this type, otherwise
-// prints all keys
-func (a *AuthCommand) ExportAuthorities(client *auth.TunClient) error {
-	var typesToExport []services.CertAuthType
-
-	// if no --type flag is given, export all types
-	if a.authType == "" {
-		typesToExport = []services.CertAuthType{services.HostCA, services.UserCA}
-	} else {
-		authType := services.CertAuthType(a.authType)
-		if err := authType.Check(); err != nil {
-			return trace.Wrap(err)
-		}
-		typesToExport = []services.CertAuthType{authType}
-	}
-	localAuthName, err := client.GetDomainName()
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	// fetch authorities via auth API (and only take local CAs, ignoring
-	// trusted ones)
-	var authorities []services.CertAuthority
-	for _, at := range typesToExport {
-		cas, err := client.GetCertAuthorities(at, a.exportPrivateKeys)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		for _, ca := range cas {
-			if ca.GetClusterName() == localAuthName {
-				authorities = append(authorities, ca)
-			}
-		}
-	}
-
-	// print:
-	for _, ca := range authorities {
-		if a.exportPrivateKeys {
-			for _, key := range ca.GetSigningKeys() {
-				fingerprint, err := sshutils.PrivateKeyFingerprint(key)
-				if err != nil {
-					return trace.Wrap(err)
-				}
-				if a.exportAuthorityFingerprint != "" && fingerprint != a.exportAuthorityFingerprint {
-					continue
-				}
-				os.Stdout.Write(key)
-				fmt.Fprintf(os.Stdout, "\n")
-			}
-		} else {
-			for _, keyBytes := range ca.GetCheckingKeys() {
-				fingerprint, err := sshutils.AuthorizedKeyFingerprint(keyBytes)
-				if err != nil {
-					return trace.Wrap(err)
-				}
-				if a.exportAuthorityFingerprint != "" && fingerprint != a.exportAuthorityFingerprint {
-					continue
-				}
-
-				// export certificates in the old 1.0 format where host and user
-				// certificate authorities were exported in the known_hosts format.
-				if a.compatVersion == "1.0" {
-					castr, err := hostCAFormat(ca, keyBytes, client)
-					if err != nil {
-						return trace.Wrap(err)
-					}
-
-					fmt.Println(castr)
-					continue
-				}
-
-				// export certificate authority in user or host ca format
-				var castr string
-				switch ca.GetType() {
-				case services.UserCA:
-					castr, err = userCAFormat(ca, keyBytes)
-				case services.HostCA:
-					castr, err = hostCAFormat(ca, keyBytes, client)
-				default:
-					return trace.BadParameter("unknown user type: %q", ca.GetType())
-				}
-				if err != nil {
-					return trace.Wrap(err)
-				}
-
-				// print the export friendly string
-				fmt.Println(castr)
-			}
-		}
-	}
-	return nil
-}
-
-// userCAFormat returns the certificate authority public key exported as a single
-// line that can be placed in ~/.ssh/authorized_keys file. The format adheres to the
-// man sshd (8) authorized_keys format, a space-separated list of: options, keytype,
-// base64-encoded key, comment.
-// For example:
-//
-//    cert-authority AAA... type=user&clustername=cluster-a
-//
-// URL encoding is used to pass the CA type and cluster name into the comment field.
-func userCAFormat(ca services.CertAuthority, keyBytes []byte) (string, error) {
-	comment := url.Values{
-		"type":        []string{string(services.UserCA)},
-		"clustername": []string{ca.GetClusterName()},
-	}
-
-	return fmt.Sprintf("cert-authority %s %s", strings.TrimSpace(string(keyBytes)), comment.Encode()), nil
-}
-
-// hostCAFormat returns the certificate authority public key exported as a single line
-// that can be placed in ~/.ssh/authorized_hosts. The format adheres to the man sshd (8)
-// authorized_hosts format, a space-separated list of: marker, hosts, key, and comment.
-// For example:
-//
-//    @cert-authority *.cluster-a ssh-rsa AAA... type=host
-//
-// URL encoding is used to pass the CA type and allowed logins into the comment field.
-func hostCAFormat(ca services.CertAuthority, keyBytes []byte, client *auth.TunClient) (string, error) {
-	comment := url.Values{
-		"type": []string{string(ca.GetType())},
-	}
-
-	roles, err := services.FetchRoles(ca.GetRoles(), client)
-	if err != nil {
-		return "", trace.Wrap(err)
-	}
-	allowedLogins, _ := roles.CheckLogins(defaults.MinCertDuration + time.Second)
-	if len(allowedLogins) > 0 {
-		comment["logins"] = allowedLogins
-	}
-
-	return fmt.Sprintf("@cert-authority *.%s %s %s",
-		ca.GetClusterName(), strings.TrimSpace(string(keyBytes)), comment.Encode()), nil
-}
-
-// GenerateKeys generates a new keypair
-func (a *AuthCommand) GenerateKeys() error {
-	keygen := native.New()
-	defer keygen.Close()
-	privBytes, pubBytes, err := keygen.GenerateKeyPair("")
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	err = ioutil.WriteFile(a.genPubPath, pubBytes, 0600)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	err = ioutil.WriteFile(a.genPrivPath, privBytes, 0600)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	fmt.Printf("wrote public key to: %v and private key to: %v\n", a.genPubPath, a.genPrivPath)
-	return nil
-}
-
-// GenerateAndSignKeys generates a new keypair and signs it for role
-func (a *AuthCommand) GenerateAndSignKeys(clusterApi *auth.TunClient) error {
-	// parse compatibility parameter
-	compatibility, err := utils.CheckCompatibilityFlag(a.compatibility)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	// generate a keypair:
-	key, err := client.NewKey()
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	// sign it and produce a cert:
-	key.Cert, err = clusterApi.GenerateUserCert(key.Pub, a.genUser, a.genTTL, compatibility)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	// write the cert+private key to the output:
-	err = client.MakeIdentityFile(a.genUser, a.output, key, a.outputFormat)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	if a.output != "" {
-		fmt.Printf("\nThe certificate has been written to %s\n", a.output)
-	}
-	return nil
 }
 
 // connectToAuthService creates a valid client connection to the auth service
@@ -688,16 +254,6 @@ func connectToAuthService(cfg *service.Config) (client *auth.TunClient, err erro
 	return client, nil
 }
 
-// validateConfig validates and updates tctl configuration
-func validateConfig(cfg *service.Config) {
-	var err error
-	// read a host UUID for this node
-	cfg.HostUUID, err = utils.ReadHostUUID(cfg.DataDir)
-	if err != nil {
-		utils.FatalError(err)
-	}
-}
-
 // applyConfig takes configuration values from the config file and applies
 // them to 'service.Config' object
 func applyConfig(ccf *CLIConfig, cfg *service.Config) error {
@@ -722,367 +278,11 @@ func applyConfig(ccf *CLIConfig, cfg *service.Config) error {
 		utils.InitLogger(utils.LoggingForCLI, logrus.DebugLevel)
 		logrus.Debugf("DEBUG loggign enabled")
 	}
-	return nil
-}
 
-// onTokenList is called to execute "tokens ls" command
-func (c *TokenCommand) List(client *auth.TunClient) error {
-	tokens, err := client.GetTokens()
+	// read a host UUID for this node
+	cfg.HostUUID, err = utils.ReadHostUUID(cfg.DataDir)
 	if err != nil {
-		return trace.Wrap(err)
+		utils.FatalError(err)
 	}
-	if len(tokens) == 0 {
-		fmt.Println("No active tokens found.")
-		return nil
-	}
-	tokensView := func() string {
-		table := goterm.NewTable(0, 10, 5, ' ', 0)
-		printHeader(table, []string{"Token", "Role", "Expiry Time (UTC)"})
-		for _, t := range tokens {
-			expiry := "never"
-			if t.Expires.Unix() > 0 {
-				expiry = t.Expires.Format(time.RFC822)
-			}
-			fmt.Fprintf(table, "%v\t%v\t%s\n", t.Token, t.Roles.String(), expiry)
-		}
-		return table.String()
-	}
-	fmt.Printf(tokensView())
 	return nil
 }
-
-// onTokenList is called to execute "tokens del" command
-func (c *TokenCommand) Del(client *auth.TunClient) error {
-	if c.token == "" {
-		return trace.Errorf("Need an argument: token")
-	}
-	if err := client.DeleteToken(c.token); err != nil {
-		return trace.Wrap(err)
-	}
-	fmt.Printf("Token %s has been deleted\n", c.token)
-	return nil
-}
-
-// Get prints one or many resources of a certain type
-func (g *GetCommand) Get(client *auth.TunClient) error {
-	collection, err := g.getCollection(client)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	switch g.format {
-	case formatText:
-		return collection.writeText(os.Stdout)
-	case formatJSON:
-		return collection.writeJSON(os.Stdout)
-	case formatYAML:
-		return collection.writeYAML(os.Stdout)
-	}
-	return trace.BadParameter("unsupported format")
-}
-
-// Create updates or insterts one or many resources
-func (u *CreateCommand) Create(client *auth.TunClient) error {
-	var reader io.ReadCloser
-	var err error
-	if u.filename != "" {
-		reader, err = utils.OpenFile(u.filename)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-	} else {
-		reader = ioutil.NopCloser(os.Stdin)
-	}
-	decoder := kyaml.NewYAMLOrJSONDecoder(reader, 32*1024)
-	count := 0
-	for {
-		var raw services.UnknownResource
-		err := decoder.Decode(&raw)
-		if err != nil {
-			if err == io.EOF {
-				if count == 0 {
-					return trace.BadParameter("no resources found, emtpy input?")
-				}
-				return nil
-			}
-			return trace.Wrap(err)
-		}
-		count += 1
-		switch raw.Kind {
-		case services.KindSAMLConnector:
-			conn, err := services.GetSAMLConnectorMarshaler().UnmarshalSAMLConnector(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := conn.CheckAndSetDefaults(); err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertSAMLConnector(conn); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created SAML connector: %v\n", conn.GetName())
-		case services.KindOIDCConnector:
-			conn, err := services.GetOIDCConnectorMarshaler().UnmarshalOIDCConnector(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertOIDCConnector(conn); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created OIDC connector: %v\n", conn.GetName())
-		case services.KindReverseTunnel:
-			tun, err := services.GetReverseTunnelMarshaler().UnmarshalReverseTunnel(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertReverseTunnel(tun); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created reverse tunnel: %v\n", tun.GetName())
-		case services.KindCertAuthority:
-			ca, err := services.GetCertAuthorityMarshaler().UnmarshalCertAuthority(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertCertAuthority(ca); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created cert authority: %v \n", ca.GetName())
-		case services.KindUser:
-			user, err := services.GetUserMarshaler().UnmarshalUser(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertUser(user); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created user: %v\n", user.GetName())
-		case services.KindRole:
-			role, err := services.GetRoleMarshaler().UnmarshalRole(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			err = role.CheckAndSetDefaults()
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertRole(role, backend.Forever); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created role: %v\n", role.GetName())
-		case services.KindNamespace:
-			ns, err := services.UnmarshalNamespace(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertNamespace(*ns); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created namespace: %v\n", ns.Metadata.Name)
-		case services.KindTrustedCluster:
-			tc, err := services.GetTrustedClusterMarshaler().Unmarshal(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.UpsertTrustedCluster(tc); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("created trusted cluster: %q\n", tc.GetName())
-		case services.KindClusterAuthPreference:
-			cap, err := services.GetAuthPreferenceMarshaler().Unmarshal(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.SetClusterAuthPreference(cap); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("updated cluster auth preferences\n")
-		case services.KindUniversalSecondFactor:
-			universalSecondFactor, err := services.GetUniversalSecondFactorMarshaler().Unmarshal(raw.Raw)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if err := client.SetUniversalSecondFactor(universalSecondFactor); err != nil {
-				return trace.Wrap(err)
-			}
-			fmt.Printf("updated utf settings\n")
-		case "":
-			return trace.BadParameter("missing resource kind")
-		default:
-			return trace.BadParameter("%q is not supported", raw.Kind)
-		}
-	}
-}
-
-// Delete deletes resource by name
-func (d *DeleteCommand) Delete(client *auth.TunClient) error {
-	if d.ref.Kind == "" {
-		return trace.BadParameter("provide full resource name to delete e.g. roles/example")
-	}
-	if d.ref.Name == "" {
-		return trace.BadParameter("provide full resource name to delete e.g. roles/example")
-	}
-
-	switch d.ref.Kind {
-	case services.KindUser:
-		if err := client.DeleteUser(d.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("user %v has been deleted\n", d.ref.Name)
-	case services.KindSAMLConnector:
-		if err := client.DeleteSAMLConnector(d.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("SAML Connector %v has been deleted\n", d.ref.Name)
-	case services.KindOIDCConnector:
-		if err := client.DeleteOIDCConnector(d.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("OIDC Connector %v has been deleted\n", d.ref.Name)
-	case services.KindReverseTunnel:
-		if err := client.DeleteReverseTunnel(d.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("reverse tunnel %v has been deleted\n", d.ref.Name)
-	case services.KindRole:
-		if err := client.DeleteRole(d.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("role %v has been deleted\n", d.ref.Name)
-	case services.KindNamespace:
-		if err := client.DeleteNamespace(d.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("namespace %v has been deleted\n", d.ref.Name)
-	case services.KindTrustedCluster:
-		if err := client.DeleteTrustedCluster(d.ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("trusted cluster %q has been deleted\n", d.ref.Name)
-	case "":
-		return trace.BadParameter("missing resource kind")
-	default:
-		return trace.BadParameter("%q is not supported", d.ref.Kind)
-	}
-
-	return nil
-}
-
-func (g *GetCommand) getCollection(client auth.ClientI) (collection, error) {
-	if g.ref.Kind == "" {
-		return nil, trace.BadParameter("specify resource to list, e.g. 'tctl get roles'")
-	}
-	switch g.ref.Kind {
-	case services.KindSAMLConnector:
-		connectors, err := client.GetSAMLConnectors(g.withSecrets)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &samlCollection{connectors: connectors}, nil
-	case services.KindOIDCConnector:
-		connectors, err := client.GetOIDCConnectors(g.withSecrets)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &oidcCollection{connectors: connectors}, nil
-	case services.KindReverseTunnel:
-		tunnels, err := client.GetReverseTunnels()
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &reverseTunnelCollection{tunnels: tunnels}, nil
-	case services.KindCertAuthority:
-		userAuthorities, err := client.GetCertAuthorities(services.UserCA, g.withSecrets)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		hostAuthorities, err := client.GetCertAuthorities(services.HostCA, g.withSecrets)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		userAuthorities = append(userAuthorities, hostAuthorities...)
-		return &authorityCollection{cas: userAuthorities}, nil
-	case services.KindUser:
-		users, err := client.GetUsers()
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &userCollection{users: users}, nil
-	case services.KindNode:
-		nodes, err := client.GetNodes(g.namespace)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &serverCollection{servers: nodes}, nil
-	case services.KindAuthServer:
-		servers, err := client.GetAuthServers()
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &serverCollection{servers: servers}, nil
-	case services.KindProxy:
-		servers, err := client.GetAuthServers()
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &serverCollection{servers: servers}, nil
-	case services.KindRole:
-		if g.ref.Name == "" {
-			roles, err := client.GetRoles()
-			if err != nil {
-				return nil, trace.Wrap(err)
-			}
-			return &roleCollection{roles: roles}, nil
-		}
-		role, err := client.GetRole(g.ref.Name)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &roleCollection{roles: []services.Role{role}}, nil
-	case services.KindNamespace:
-		if g.ref.Name == "" {
-			namespaces, err := client.GetNamespaces()
-			if err != nil {
-				return nil, trace.Wrap(err)
-			}
-			return &namespaceCollection{namespaces: namespaces}, nil
-		}
-		ns, err := client.GetNamespace(g.ref.Name)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &namespaceCollection{namespaces: []services.Namespace{*ns}}, nil
-	case services.KindTrustedCluster:
-		if g.ref.Name == "" {
-			trustedClusters, err := client.GetTrustedClusters()
-			if err != nil {
-				return nil, trace.Wrap(err)
-			}
-			return &trustedClusterCollection{trustedClusters: trustedClusters}, nil
-		}
-		trustedCluster, err := client.GetTrustedCluster(g.ref.Name)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &trustedClusterCollection{trustedClusters: []services.TrustedCluster{trustedCluster}}, nil
-	case services.KindClusterAuthPreference:
-		cap, err := client.GetClusterAuthPreference()
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &authPreferenceCollection{AuthPreference: cap}, nil
-	case services.KindUniversalSecondFactor:
-		universalSecondFactor, err := client.GetUniversalSecondFactor()
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &universalSecondFactorCollection{UniversalSecondFactor: universalSecondFactor}, nil
-	}
-
-	return nil, trace.BadParameter("'%v' is not supported", g.ref.Kind)
-}
-
-const (
-	formatYAML = "yaml"
-	formatText = "text"
-	formatJSON = "json"
-)
