+++ tool/tctl/common/user_command.go
@@ -1,0 +1,120 @@
+/*
+Copyright 2015-2017 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package common
+
+import (
+	"fmt"
+	"net"
+	"os"
+	"strconv"
+	"strings"
+
+	"github.com/gravitational/teleport/lib/auth"
+	"github.com/gravitational/teleport/lib/defaults"
+	"github.com/gravitational/teleport/lib/service"
+	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/web"
+	"github.com/gravitational/trace"
+)
+
+// UserCommand implements `tctl users` set of commands
+type UserCommand struct {
+	config        *service.Config
+	login         string
+	allowedLogins string
+	roles         string
+	identities    []string
+}
+
+// Add creates a new sign-up token and prints a token URL to stdout.
+// A user is not created until he visits the sign-up URL and completes the process
+func (u *UserCommand) Add(client *auth.TunClient) error {
+	// if no local logins were specified, default to 'login'
+	if u.allowedLogins == "" {
+		u.allowedLogins = u.login
+	}
+	user := services.UserV1{
+		Name:          u.login,
+		AllowedLogins: strings.Split(u.allowedLogins, ","),
+	}
+	token, err := client.CreateSignupToken(user)
+	if err != nil {
+		return err
+	}
+	proxies, err := client.GetProxies()
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	hostname := "teleport-proxy"
+	if len(proxies) == 0 {
+		fmt.Printf("\x1b[1mWARNING\x1b[0m: this Teleport cluster does not have any proxy servers online.\nYou need to start some to be able to login.\n\n")
+	} else {
+		hostname = proxies[0].GetHostname()
+	}
+
+	// try to auto-suggest the activation link
+	_, proxyPort, err := net.SplitHostPort(u.config.Proxy.WebAddr.Addr)
+	if err != nil {
+		proxyPort = strconv.Itoa(defaults.HTTPListenPort)
+	}
+	url := web.CreateSignupLink(net.JoinHostPort(hostname, proxyPort), token)
+	fmt.Printf("Signup token has been created and is valid for %v seconds. Share this URL with the user:\n%v\n\nNOTE: make sure '%s' is accessible!\n", defaults.MaxSignupTokenTTL.Seconds(), url, hostname)
+	return nil
+}
+
+// Update updates existing user
+func (u *UserCommand) Update(client *auth.TunClient) error {
+	user, err := client.GetUser(u.login)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	roles := strings.Split(u.roles, ",")
+	for _, role := range roles {
+		if _, err := client.GetRole(role); err != nil {
+			return trace.Wrap(err)
+		}
+	}
+	user.SetRoles(roles)
+	if err := client.UpsertUser(user); err != nil {
+		return trace.Wrap(err)
+	}
+	fmt.Printf("%v has been updated with roles %v\n", user.GetName(), strings.Join(user.GetRoles(), ","))
+	return nil
+}
+
+// List prints all existing user accounts
+func (u *UserCommand) List(client *auth.TunClient) error {
+	users, err := client.GetUsers()
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	coll := &userCollection{users: users}
+	coll.writeText(os.Stdout)
+	return nil
+}
+
+// Delete deletes teleport user(s). User IDs are passed as a comma-separated
+// list in UserCommand.login
+func (u *UserCommand) Delete(client *auth.TunClient) error {
+	for _, l := range strings.Split(u.login, ",") {
+		if err := client.DeleteUser(l); err != nil {
+			return trace.Wrap(err)
+		}
+		fmt.Printf("User '%v' has been deleted\n", l)
+	}
+	return nil
+}
